+++ lib/modules/security/security_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: security
-// Updated at: 2026-04-01 23:23:49
-// Commit: test(security): add unit tests for security business logic
+// Updated at: 2026-04-02 16:12:53
+// Commit: test(security): implement widget integration tests for security
 
 class securityComponentState {
-  final String commitRef = '37';
-  final String lastMessage = 'test(security): add unit tests for security business logic';
-  final DateTime timestamp = DateTime.parse('2026-04-01T23:23:49+05:00');
+  final String commitRef = '67';
+  final String lastMessage = 'test(security): implement widget integration tests for security';
+  final DateTime timestamp = DateTime.parse('2026-04-02T16:12:53+05:00');
 }